requirements.txt
apscheduler>=3.10.0
fastapi>=0.104.0
uvicorn>=0.24.0
pyjwt>=2.8.0
python-multipart>=0.0.6
psycopg[binary]>=3.1.0
redis>=5.0.0
